[feature/4] Adicionando validações

+++ src/main/java/puc/tcc/logistics/services/impl/SupplierServiceImpl.java
@@ -2,12 +2,15 @@ package puc.tcc.logistics.services.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import puc.tcc.logistics.exception.LogisticsException;
 import puc.tcc.logistics.mapper.SupplierMapper;
 import puc.tcc.logistics.persistence.repositories.SupplierRepository;
 import puc.tcc.logistics.resources.supplier.SupplierRequest;
 import puc.tcc.logistics.resources.supplier.SupplierResponse;
 import puc.tcc.logistics.services.SupplierService;
+import puc.tcc.logistics.utils.CNPJUtils;
 
 import javax.transaction.Transactional;
 import java.util.ArrayList;
@@ -26,10 +29,14 @@ public class SupplierServiceImpl implements SupplierService {
 
     @Override
     @Transactional
-    public SupplierResponse saveOrUpdate(final SupplierRequest supplierRequest) {
-        var model = supplierMapper.toModel(supplierRequest);
-        model = supplierRepository.save(model);
-        return supplierMapper.toResponse(model);
+    public SupplierResponse saveOrUpdate(final SupplierRequest supplierRequest) throws LogisticsException {
+        if (CNPJUtils.validate(supplierRequest.getCnpj())) {
+            var model = supplierMapper.toModel(supplierRequest);
+            model = supplierRepository.save(model);
+            return supplierMapper.toResponse(model);
+        } else {
+            throw new LogisticsException(HttpStatus.BAD_REQUEST, "cnpj", "CNPJ inválido!");
+        }
     }
 
     @Override
@@ -48,7 +55,11 @@ public class SupplierServiceImpl implements SupplierService {
 
     @Override
     @Transactional
-    public void delete(final Long id){
+    public void delete(final Long id) throws LogisticsException {
+        var supplier = supplierRepository.findById(id);
+        if(supplier.isEmpty()){
+            throw new LogisticsException(HttpStatus.NOT_FOUND, "supplier", "Fornecedor não existe!");
+        }
         supplierRepository.deleteById(id);
     }
 }
